@@ -4,7 +4,7 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.1",
         "ext-mbstring": "*",
         "doctrine/collections": "^1.6",
         "spiral/cycle-bridge": "^1.0",
